% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slick.R
\name{slick}
\alias{slick}
\title{Slick carousel}
\usage{
slick(dots = TRUE, infinite = TRUE, slidesToShow = 1, slidesToScroll = 1, ui)
}
\arguments{
\item{dots}{Whether to show dots.}

\item{infinite}{infinite scrolling}

\item{slidesToShow}{number of slides to show in each window}

\item{slidesToScroll}{number of slides to scroll each time}

\item{ui}{Elements to show in the slider.}
}
\value{

}
\description{
Slick carousel
}
\examples{
slick(ui = shiny::tagList(shiny::div("content1"), shiny::div("content2")))
}
