% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slick.R
\name{slick}
\alias{slick}
\title{Slick carousel}
\usage{
slick(...)
}
\arguments{
\item{...}{Div elements to show in the slider.}
}
\value{

}
\description{
Slick carousel
}
\examples{
slick(shiny::div("content1"), shiny::div("content2"))
}
